requirements.txt
hydra-core
pytorch-lightning
torch
transformers
wandb